requirements.txt
flask
flask-socketio
requests
pytest
pytest-playwright
